chore: refactored folder structure for better readabilty, and alignment with good development practices
