+++ app/globals.css
@@ -20,6 +20,12 @@ body {
   font-family: Arial, Helvetica, sans-serif;
 }
 
+.light {
+  color: var(--foreground);
+  --background: #ededed;
+  --foreground: #0a0a0a;
+}
+
 .dark {
   --background: #1a202c;
   --foreground: #f7fafc;
