+++ app/page.tsx
@@ -1,5 +1,5 @@
 import Image from "next/image";
-import Hero from "./components/Hero";
+import Hero from "../components/Hero";
 import { Analytics } from "@vercel/analytics/next";
 
 export default function Home() {
