+++ components/Hero.tsx
