+++ components/quiz/QuizContent.tsx
