+++ components/quiz/QuizProgress.tsx
