+++ components/quiz/SubdomainProgress.tsx
