+++ components/quiz/types.ts
