+++ questions/cloud-practitioner/1-cloud-concepts/1.2.ts
