+++ questions/cloud-practitioner/2-security-and-compliance/2.1.ts
