+++ questions/cloud-practitioner/2-security-and-compliance/2.4.ts
