+++ questions/cloud-practitioner/3-technology/3.1.ts
