+++ questions/cloud-practitioner/3-technology/3.2.ts
