+++ questions/cloud-practitioner/3-technology/3.3.ts
