+++ questions/cloud-practitioner/3-technology/3.4.ts
