+++ questions/sysops-administrator-associate/1-monitoring-logging/1.1.ts
