+++ questions/sysops-administrator-associate/1-monitoring-logging/1.2.ts
