+++ questions/sysops-administrator-associate/3-deployment-provisioning-automation/3.2.ts
