+++ questions/sysops-administrator-associate/6-cost-performance-optimization/6.1.ts
