+++ tailwind.config.ts
@@ -11,3 +11,6 @@ export default {
   },
   plugins: [],
 } satisfies Config;
+
+
+
